Add files via upload
FrameLayout:
Android Framelayout is a ViewGroup subclass that is used to specify the position of multiple views placed on top of each other to represent a single view screen. Generally, we can say FrameLayout simply blocks a particular area on the screen to display a single view because it can be difficult to display single views at a specific area on the screen without overlapping each other. Here, all the child views or elements are added in stack format means the most recently added child will be shown on the top of the screen. But, we can add multiple children’s views and control their positions only by using gravity attributes in FrameLayout.
Syntax:
<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
android:layout_width="match_parent"
android:layout_height="match_parent"
android:orientation="vertical">
// Add items or widgets here
</FrameLayout>
Attributes of Frame Layout:
1. android:id= This is the unique id which identifies the layout in the R.java file.
2.android:foreground= Foreground defines the drawable to draw over the content and this may be a color value. 
On applying android:foregroud feature, all the views taken in the framelayout will goes to the background and the framelayout comes in the foreground i.e. over the views. We can set the @drawable/image_name or the color in the foreground.
android:foreground="#0f0"
3. android:visibility= This determine whether to make the view visible, invisible or gone.
visible – the view is present and also visible invisible – The view is present but not visible gone – The view is neither present nor visible